Renaming test to match objective
